refactor: test about cmp subcommand was modularized
