+++ tests/cli/cmp_subcommand/mod.rs
@@ -1,0 +1,11 @@
+/*
+ *  Quick Test: CLI for stress testing in competitive programming
+ *  Copyright (C) 2021 - Luis Miguel Báez
+ *  License: MIT (See the LICENSE file in the repository root directory)
+ */
+
+pub mod codes;
+pub mod cpp_python;
+pub mod cpp_ce;
+pub mod cpp_rte;
+pub mod python_rte;
